Replace spec.prepare by spec.middleware
Explanation: some middlewares (i.e. bicing) declare functions as
l-values of the returned functions. If we use regular require we can
modify the scope of the returned function only; not the lvalues. So, if
a function (different from the returned one) uses, say, xml, it gives
an error with the old implementation.